Moving addAppender to DiagnosticsAppender, using config and reflection to instantiate
