+++ samza-core/src/main/scala/org/apache/samza/config/JobConfig.scala
@@ -91,6 +91,10 @@ object JobConfig {
   // Enables diagnostic appender for logging exception events
   val JOB_DIAGNOSTICS_ENABLED = "job.diagnostics.enabled"
 
+  // Specify DiagnosticAppender class
+  val DIAGNOSTICS_APPENDER_CLASS = "job.diagnostics.appender.class"
+
+
   implicit def Config2Job(config: Config) = new JobConfig(config)
 
   /**
@@ -191,4 +195,6 @@ class JobConfig(config: Config) extends ScalaMapConfig(config) with Logging {
   def getLoggedStorePath = getOption(JobConfig.JOB_LOGGED_STORE_BASE_DIR)
 
   def getDiagnosticsEnabled = { getBoolean(JobConfig.JOB_DIAGNOSTICS_ENABLED, false) }
+
+  def getDiagnosticsAppenderClass = { getOption(JobConfig.DIAGNOSTICS_APPENDER_CLASS) }
 }
